HOR-88: regenerar dao http://jira.uji.es/browse/HOR-88
Regenerado el DAO y arreglados algunos tipos de datos de BigDecimal a Long

git-svn-id: http://jira.uji.es/repos/uji/HOR/uji-hor-base/trunk@11441 e6edf6fb-f266-4316-afb4-e53d95876a76

+++ src/main/java/es/uji/apps/hor/dao/CursosDAODatabaseImpl.java
@@ -1,6 +1,5 @@
 package es.uji.apps.hor.dao;
 
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -31,7 +30,7 @@ public class CursosDAODatabaseImpl extends BaseDAODatabaseImpl implements Cursos
 
         for (Tuple tuple : listaCursosTuples)
         {
-            BigDecimal idCurso = tuple.get(item.cursoId);
+            Long idCurso = tuple.get(item.cursoId);
 
             cursos.add(new Curso(idCurso.longValue()));
         }
